face/requirements.txt
fastapi
uvicorn
face_recognition
opencv-python
numpy
pillow
